handle large data

+++ ccm.h
@@ -13,7 +13,7 @@ typedef struct {
 unsigned char* ccm_encrypt(int*, ccm_t*);
 unsigned char* ccm_decrypt(int*, ccm_decrypt_t*);
 void print_block(unsigned char*);
-void format(ccm_t*, unsigned char **, int, unsigned char flags);
+void format(ccm_t*, unsigned char **, unsigned long, unsigned char flags);
 void gen_ctr(unsigned char **, int, unsigned char *, unsigned long, unsigned char);
 
 /* error handling */
